Refazendo o widget buttons de acordo com prototipo
